Feature/style UI (#5)
* Add get grouped activities method

* Move activity items to their own component

* Add content and styling

* Activity filters styling

Add react calendar package. Style the react calendar. Style the filters.

* Clean up components

* Error testing

Install react toastify. Add test error component to test bad server responses

* Api Error handling

Add in testing page for errors for a visual check of the responses. Add in HistoryRouter so this can be accessed from the agent.ts since it is NOT a template file.

* Handle validation errors and bad request

* Add server error interface

* Add commonStore

* Use common store

* Add server error template

* Pass data to server error template, add route

* Add functionality for bad guid button